+++ package.json
@@ -49,6 +49,7 @@
     "@types/react-helmet": "^6.1.1",
     "prettier": "2.0.5",
     "tailwindcss": "^1.6.2",
+    "ts-node": "^9.1.1",
     "typescript": "^4.1.3"
   },
   "homepage": "https://github.com/renyuanz/leonids#readme",
